Move s.u.FromString to s.u.CommandLineParser.FromString.
There are many competing ways to parse things. `FromString` is
tailored for command-line parsing. As such we should not expose it
with a general name like `scala.util.FromString`. Instead, we move
it inside `CommandLineParser` to make it clear that it is meant for
command-line parsing.
