+++ tests/neg/main-functions2.scala
@@ -8,6 +8,6 @@ class Foo {
 
 @main def g(x: Int*)(y: Int*) = () // error: varargs parameter of @main method must come last
 
-@main def h[T: util.FromString](x: T) = () // error: @main method cannot have type parameters
+@main def h[T: util.CommandLineParser.FromString](x: T) = () // error: @main method cannot have type parameters
 
 @main def i(x: Int)(using Foo) = () // error: @main method cannot have implicit parameters
