+++ tests/run/decorators/main.scala
@@ -5,7 +5,7 @@ import collection.mutable
  */
 class main extends EntryPoint.Annotation:
 
-  type ArgumentParser[T] = util.FromString[T]
+  type ArgumentParser[T] = util.CommandLineParser.FromString[T]
   type EntryPointResult  = Unit
 
   inline def wrapperName(entryPointName: String): String =
